save bugfix when ordner nicht da

+++ src/graphics/Frame.java
@@ -134,9 +134,10 @@ public class Frame extends JFrame {
 
 	private void deleteAll() {
 		File tmpSaveFolder = new File(System.getenv("APPDATA") + "\\tetris-n-run\\levelSaves\\tmpSaves");
-		for (File f : tmpSaveFolder.listFiles()) {
-			f.delete();
-		}
+		if (tmpSaveFolder.exists())
+			for (File f : tmpSaveFolder.listFiles()) {
+				f.delete();
+			}
 	}
 
 	// Level to level
